refactor: remove phasor common utils entry
