+++ packages/phasor/src/grid.ts
@@ -1,12 +1,33 @@
 import type { Element } from './elements/index.js';
-import type { IBound } from './consts.js';
-import {
-  compare,
-  getGridIndex,
-  isOverlap,
-  isPointIn,
-  rangeFromBound,
-} from './utils.js';
+import { GRID_SIZE, IBound } from './consts.js';
+
+function isPointIn(a: IBound, x: number, y: number): boolean {
+  return a.x < x && x <= a.x + a.w && a.y < y && y <= a.y + a.h;
+}
+
+function isOverlap(a: IBound, b: IBound): boolean {
+  return (
+    a.x < b.x + b.w && a.x + a.w > b.x && a.y < b.y + b.h && a.y + a.h > b.y
+  );
+}
+
+function getGridIndex(val: number) {
+  return Math.ceil(val / GRID_SIZE) - 1;
+}
+
+function rangeFromBound(a: IBound): number[] {
+  const minRow = getGridIndex(a.x);
+  const maxRow = getGridIndex(a.x + a.w);
+  const minCol = getGridIndex(a.y);
+  const maxCol = getGridIndex(a.y + a.h);
+  return [minRow, maxRow, minCol, maxCol];
+}
+
+function compare(a: Element, b: Element): number {
+  if (a.index > b.index) return 1;
+  else if (a.index < b.index) return -1;
+  return a.id > b.id ? 1 : -1;
+}
 
 export class GridManager {
   private _grids: Map<string, Set<Element>> = new Map();
