+++ packages/phasor/src/utils/debug-utils.ts
@@ -1,5 +1,5 @@
-import { DebugElement, type Element } from './elements/index.js';
-import type { Renderer } from './renderer.js';
+import { DebugElement, type Element } from '../elements/index.js';
+import type { Renderer } from '../renderer.js';
 
 const DEBUG_ELEMENT_MAX = 150;
 const DEBUG_ELEMEMNT_MIN = 120;
